# frontend/src/lib/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing blips, technologies, users, and user technologies
  title: Tech Radar API
  contact: {}
  version: "1.0"
paths:
...
  /technologies:
    get:
      description: Get all technologies
      tags:
        - technologies
      summary: Get all technologies
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/handlers.Technology"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    post:
      description: Create a new technology (automatically creates a blip)
      tags:
        - technologies
      summary: Create a technology
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.CreateTechnologyRequest"
        description: Technology data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Technology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/technologies/by-name/{name}":
    get:
      description: Get technology by name
      tags:
        - technologies
      summary: Get a technology by name
      parameters:
        - description: Technology name
          name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Technology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/technologies/by-quadrant/{quadrant_id}":
    get:
      description: Get all technologies in a quadrant
      tags:
        - technologies
      summary: Get technologies by quadrant
      parameters:
        - description: Quadrant ID
          name: quadrant_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/handlers.Technology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/technologies/{id}":
    get:
      description: Get technology by ID
      tags:
        - technologies
      summary: Get a technology
      parameters:
        - description: Technology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Technology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    put:
      description: Update technology by ID
      tags:
        - technologies
      summary: Update a technology
      parameters:
        - description: Technology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.UpdateTechnologyRequest"
        description: Technology data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Technology"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    delete:
      description: Delete technology by ID
      tags:
        - technologies
      summary: Delete a technology
      parameters:
        - description: Technology ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
...
servers:
  - url: //localhost:8080
components:
  schemas:
...
    handlers.CreateTechnologyRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        quadrant_id:
          type: integer
...
    handlers.Error:
      type: object
      properties:
        message:
          type: string
    handlers.Technology:
      type: object
      properties:
        blip_id:
          type: integer
          example: 1
        created_at:
          type: string
          example: 2026-04-05T12:00:00Z
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: AI Platform Insights
        quadrant_id:
          type: integer
          example: 2
        updated_at:
          type: string
          example: 2026-04-05T12:00:00Z
...
    handlers.UpdateTechnologyRequest:
      type: object
      properties:
        blip_id:
          type: integer
        name:
          type: string
        quadrant_id:
          type: integer
...